+++ src/export/cast.rs
@@ -39,6 +39,7 @@ mod tests {
             title: "t".into(),
             events: vec![(0.1, "hi".into()), (0.2, "\r\n".into())],
             captions: vec![],
+            focuses: vec![],
             duration: 0.7,
         };
         let cast = to_cast(&rec).unwrap();